backend/requirements.txt
alembic==1.17.0
azure-core==1.36.0
azure-storage-blob==12.24.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
Flask==3.0.2
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==21.2.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==8.4.0
packaging==24.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.29
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.0.1
WTForms==3.0.1
zipp==3.15.0
